(#1) feat : eslint prettier husky 설정 DETAIL : --save-dev로 설치 REASON : code formatting
