+++ .prettierrc.json
@@ -1,0 +1,5 @@
+{
+  "printWidth": 100,
+  "singleQuote": true,
+  "arrowParens": "avoid"
+}
