+++ src/App.jsx
@@ -1,0 +1,5 @@
+function App() {
+  return <>app</>;
+}
+
+export default App;
